1 Player against animals is a Go!
*** TODO:
	1) Add crafting to 1 Player CTB.
	2) Fix configurator for number of i) cows, ii) chickens, iii) sheep.
	3) Create new PolycraftAnimal entities in spreadsheets so we can change
vulnerability and speed in future versions.
	4) Add vulnerability and speed to configurator.

+++ src/main/java/edu/utd/minecraft/mod/polycraft/experiment/feature/ExperimentFeature.java
@@ -27,8 +27,8 @@ public abstract class ExperimentFeature {
 	
 	public abstract void build(World world);
 	
-	public long[] getPositionArray() {
-		return new long[] {xPos, yPos, zPos};
+	public int[] getPositionArray() {
+		return new int[] {xPos, yPos, zPos};
 	}
 	
 	protected class ExperimentFeatureListDeserializer implements JsonDeserializer<ArrayList<ExperimentFeature>>{
